Changes to the image masking
Previous versions used a full size image mask to exclude regions outside the mask from being considered for Canny edge detection. Now, the code extracts the mask region and operates on that. This means that the GRAY and THRESH images are not necessarily full size hence the image processing may be a lot faster. The EDGES image is re-assembled for compatibility but there is a new call readSmallEDGES() which returns the small image for  contour detection. Contour co-ordinates must be adjusted using the mask origin which is returned by getMaskOffsets()